test: support static addresses for http endpoints
